web/requirements.txt
Flask
flask_restful
pymongo
bcrypt
numpy 
tensorflow
requests
keras
pillow
